moved unstructured dependency to worker
