+++ redbox-core/pyproject.toml
@@ -19,10 +19,7 @@ langchain_openai = "^0.1.9"
 tiktoken = "^0.7.0"
 boto3 = "^1.34.136"
 pydantic-settings = "^2.3.4"
-torch = "2.2.2"
 langchain-elasticsearch = "^0.2.2"
-unstructured = {version = "<0.14.9", extras = ["all-docs"]}
-unstructured-inference = "^0.7.36"
 pytest-dotenv = "^0.5.2"
 kneed = "^0.8.5"
