+++ worker/pyproject.toml
@@ -12,7 +12,7 @@ readme = "../README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.12,<3.13"
-redbox = {path="../redbox-core", develop=true}
+redbox = {git = "https://github.com/i-dot-ai/redbox.git", rev = "1f9e018", subdirectory = "redbox-core"}
 langchain = "^0.2.6"
 langchain-elasticsearch = "^0.2.2"
 elasticsearch = "^8.14.0"
@@ -21,6 +21,10 @@ pillow-heif = "^0.16.0"
 faststream = {extras = ["redis"], version = "<0.5.0"}
 nltk = "^3.8.1"
 pikepdf = "<9.0.0"
+torch = "2.2.2"
+unstructured = {version = "<0.14.9", extras = ["all-docs"]}
+unstructured-inference = "^0.7.36"
+
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.2.2"
